Rename `sync::sync` to `sync::blocking`
